+++ qiskit/transpiler/passes/__init__.py
@@ -13,3 +13,4 @@ from .mapping.check_map import CheckMap
 from .mapping.basic_mapper import BasicMapper
 from .mapping.direction_mapper import DirectionMapper
 from .mapping.unroller import Unroller
+from .mapping.lookahead_mapper import LookaheadMapper