+++ test/python/test_reordering.py
@@ -50,8 +50,6 @@ class TestBitReordering(QiskitTestCase):
         threshold = 0.1 * shots
         self.assertDictAlmostEqual(counts_real, counts_sim, threshold)
 
-    @unittest.skip("Temporary skipping")
-    # skipping temporarily due to mapping wire fragment bug.
     @slow_test
     @requires_qe_access
     def test_multi_register_reordering(self, qe_token, qe_url):
@@ -90,7 +88,7 @@ class TestBitReordering(QiskitTestCase):
         self.assertDictAlmostEqual(counts_real, counts_sim, threshold)
 
     def _get_backends(self, qe_token, qe_url):
-        sim_backend = qiskit.Aer.get_backend('qasm_simulator_py')
+        sim_backend = qiskit.Simulators.get_backend('qasm_simulator')
         try:
             qiskit.IBMQ.enable_account(qe_token, qe_url)
             real_backends = qiskit.IBMQ.backends(simulator=False)